last update

+++ app/src/main/java/uk/org/websolution/trader_org_tool/notes/NoteViewHolder.java
@@ -1,5 +1,9 @@
-package uk.org.websolution.geeknotepad;
+package uk.org.websolution.trader_org_tool.notes;
 
+import android.app.AlertDialog;
+import android.content.res.ColorStateList;
+import android.graphics.Color;
+import android.os.Build;
 import android.view.LayoutInflater;
 import android.view.MenuItem;
 import android.view.ViewGroup;
@@ -7,9 +11,12 @@ import android.widget.TextView;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
+import androidx.annotation.RequiresApi;
 import androidx.appcompat.widget.PopupMenu;
 import androidx.cardview.widget.CardView;
 import androidx.recyclerview.widget.RecyclerView;
+
+import uk.org.websolution.trader_org_tool.R;
 
 public class NoteViewHolder extends RecyclerView.ViewHolder implements MenuItem.OnMenuItemClickListener, PopupMenu.OnMenuItemClickListener {
     private final TextView textViewTitle;
@@ -31,9 +38,12 @@ public class NoteViewHolder extends RecyclerView.ViewHolder implements MenuItem.
         });
     }
 
+    @RequiresApi(api = Build.VERSION_CODES.O)
     public void bind(NoteEntity noteEntity) {
         this.noteEntity = noteEntity;
-        cardView.setCardBackgroundColor(noteEntity.getColour());
+        //cardView.setBackgroundTintList(ColorStateList.valueOf(Color.argb(95, 52, 73, 85)));
+        cardView.setCardBackgroundColor(ColorStateList.valueOf(Color.argb(95, 52, 73, 85)));;
+        //cardView.setBackgroundTintList(ColorStateList.valueOf(Color.parseColor("#344955")));
         textViewTitle.setText(noteEntity.getTitle());
         textViewDate.setText(noteEntity.getDate());
         itemView.setOnLongClickListener(v -> {
@@ -50,7 +60,17 @@ public class NoteViewHolder extends RecyclerView.ViewHolder implements MenuItem.
         if (item.getItemId() == R.id.edit_note) {
             clickListener.onEditClicked(noteEntity);
         } else if (item.getItemId() == R.id.delete_note) {
-            clickListener.onDeleteClicked(noteEntity);
+            new AlertDialog.Builder(cardView.getContext())
+                    .setTitle(R.string.warning)
+                    .setMessage(R.string.warning_message)
+                    .setCancelable(true)
+                    .setPositiveButton(R.string.yes, (d,i) ->{
+                        clickListener.onDeleteClicked(noteEntity);
+                    })
+                    .setNegativeButton(R.string.no, (d,i) ->{
+                        return;
+                    })
+                    .show();
         }
         return false;
     }
